+++ controllers/contacts.js
@@ -1,62 +1,77 @@
 const Joi = require("joi")
-const path = require("path")
+const path = require("path");
+const { httpError} = require(path.resolve(__dirname, "../helpers"))
 const {
     listContacts,
     getContactById,
     addContact,
     updateContact,
     removeContact,
-  } = require(path.resolve(__dirname, "../models/contacts"));
+} = require(path.resolve(__dirname, "../models/contacts"));
 
-  const { httpError, ctrlWrapper } = require(path.resolve(__dirname, "../helpers"))
-  
-  const addSchema = Joi.object({
-    name: Joi.string()
-    .trim()
-    .required(),
-    email: Joi.string()
-    .trim()
-    .email()
-    .required(),
-    phone: Joi.string()
-    .trim()
-    .regex(/^\+\d{1}\d{8,15}$/)
-    .messages({
-      'string.pattern.base': 'Phone number must start with a plus sign (+) and have 9 to 16 digits.',
-    })
-    .required()
-  });
+const addSchema = Joi.object({
+  name: Joi.string()
+  .trim()
+  .required(),
+  email: Joi.string()
+  .trim()
+  .email()
+  .required(),
+  phone: Joi.string()
+  .trim()
+  .regex(/^\+\d{1}\d{8,15}$/)
+  .messages({
+    'string.pattern.base': 'Phone number must start with a plus sign (+) and have 9 to 16 digits.',
+  })
+  .required()
+});
 
-  const getAll = async (req, res) => {
-      const contacts = await listContacts();
-      if(!contacts) {
-        throw httpError(404, "Not found")
+  const getAll =  async (req, res, next) => {
+    try {
+        const contacts = await listContacts();
+        if(!contacts) {
+          throw httpError(404, "Not found")
+        }
+        res.status(200).json( contacts );
+        
+      } catch (error) {
+        next(error)
       }
-      res.status(200).json( contacts );
-  };
-
-  const getById = async (req, res) => {
-    const contactId = req.params.contactId; 
-    const contactById = await getContactById(contactId);
-    console.log('contactById: ', contactById)
-    if(!contactById) {
-      throw httpError(404, "Not found")
-    }
-    res.status(200).json( contactById );
   }
 
-  const add = async (req, res) => {
-    const { error } = addSchema.validate(req.body);
-    if( error) {
-      throw httpError(400, error.message)
-    }
-  const newContact = await addContact(req.body); 
+  const getById = async (req, res, next) => {
+    try {
+    const contactId = req.params.contactId; // Получаем значение contactId из параметров запроса
+    const contactById = await getContactById(contactId);
+    if(!contactById) {
+        throw httpError(404, "Not found")
+      }
+      res.status(200).json( contactById );
+    
+      } catch (error) {
+        next(error)
+      }
+  }
 
+  const add = async (req, res, next) => {
+    try {
+        const {error} = addSchema.validate(req.body)
+        if(error) {
+          throw httpError(400, error.message)
+        }
+        const { name, email, phone } = req.body; // Извлекаем данные из тела запроса
+    // console.log(req.body)
+    const newContact = await addContact({ name, email, phone }); // Передаем параметр contact
+    
   res.status(201).json( newContact );
+} catch (error) {
+  next(error)
 }
+  }
 
-const deleteById = async (req, res) => {
-      const contactId = req.params.contactId;
+const deleteById = async (req, res, next) => {
+  try {
+      const contactId = req.params.contactId; // Получаем значение contactId из параметров запроса
       const removedById = await removeContact(contactId);
   
       if(!removedById) {
@@ -65,32 +80,48 @@ const deleteById = async (req, res) => {
       res.json({
         message: "Delete success"
       })  
-  }
-
-  const updateContactById = async (req, res) => {
-    const { error } = addSchema.validate(req.body);
-    if( error) {
-      throw httpError(400, error.message)
+    } catch (error) {
+      next(error)
     }
-    const contactId = req.params.contactId; 
-    const { name, email, phone } = req.body; 
+}
+
+  const updateContactById = async (req, res, next) => {
+    try {
+        const {error} = addSchema.validate(req.body)
+        if(error) {
+          throw httpError(400, error.message)
+        }
   
-    const newContact = { name, email, phone };
+    const contactId = req.params.contactId; // Получаем значение contactId из параметров запроса
+    const { name, email, phone } = req.body; // Извлекаем данные из тела запроса
   
-    const updatedContact = await updateContact(contactId, newContact);
+    const newContact = { name, email, phone }; // Создаем объект newContact
+  
+    const updatedContact = await updateContact(contactId, newContact); // Передаем параметры contactId и newContact
     if(!updatedContact) {
-      throw httpError(404, "Not found")
-    }
+        throw httpError(404, "Not found")
+      }
+    
+      res.json( updatedContact );
   
-    res.json( updatedContact );
+    } catch (error) {
+        next(error)
+      }
   }
 
-  
   module.exports = {
-    getAll: ctrlWrapper(getAll),
-    getById: ctrlWrapper(getById),
-    add: ctrlWrapper(add),
-    deleteById: ctrlWrapper(deleteById),
-    updateContactById: ctrlWrapper(updateContactById),
+    getAll,
+    getById,
+    add,
+    deleteById,
+    updateContactById,
   }
+  
+  // module.exports = {
+  //   getAll: ctrlWrapper(getAll),
+  //   getById: ctrlWrapper(getById),
+  //   add: ctrlWrapper(add),
+  //   deleteById: ctrlWrapper(deleteById),
+  //   updateContactById: ctrlWrapper(updateContactById),
+  // }
