--- controllers/index.js
@@ -1,6 +1,0 @@
-const path = require("path")
-const contacts = require(path.resolve(__dirname, "./contacts"));
-
-module.exports = {
-    contacts,
-}
